# Add warnings + error ignoring for index lookup failures