--- a/src/binder/bind/copy/bind_copy_from.cpp
+++ b/src/binder/bind/copy/bind_copy_from.cpp
@@ -87,18 +87,13 @@ std::unique_ptr<BoundStatement> Binder::bindCopyNodeFrom(const Statement& statem
         expectedColumnTypes);
     auto boundSource = bindScanSource(copyStatement.getSource(),
         copyStatement.getParsingOptionsRef(), expectedColumnNames, expectedColumnTypes);
-    expression_vector warningDataExprs;
+    expression_vector warningDataExprs = boundSource->getWarningColumns();
     if (boundSource->type == ScanSourceType::FILE) {
         auto& source = boundSource->constCast<BoundTableScanSource>();
         auto bindData = source.info.bindData->constPtrCast<ScanBindData>();
         if (copyStatement.byColumn() && bindData->config.fileTypeInfo.fileType != FileType::NPY) {
             throw BinderException(stringFormat("Copy by column with {} file type is not supported.",
                 bindData->config.fileTypeInfo.fileTypeStr));
-        }
-
-        for (column_id_t i = source.info.bindData->numWarningDataColumns; i >= 1; --i) {
-            KU_ASSERT(i < source.info.columns.size());
-            warningDataExprs.push_back(source.info.columns[source.info.columns.size() - i]);
         }
     }
     expression_vector columns;
@@ -132,19 +127,8 @@ std::unique_ptr<BoundStatement> Binder::bindCopyRelFrom(const parser::Statement&
         expectedColumnTypes, clientContext);
     auto boundSource = bindScanSource(copyStatement.getSource(),
         copyStatement.getParsingOptionsRef(), expectedColumnNames, expectedColumnTypes);
-    expression_vector warningDataExprs;
-    bool ignoreErrors = CopyConstants::DEFAULT_IGNORE_ERRORS;
-    if (boundSource->type == ScanSourceType::FILE) {
-        auto& source = boundSource->constCast<BoundTableScanSource>();
-        auto bindData = source.info.bindData->constPtrCast<ScanBindData>();
-        for (column_id_t i = bindData->numWarningDataColumns; i >= 1; --i) {
-            KU_ASSERT(i < source.info.columns.size());
-            warningDataExprs.push_back(source.info.columns[source.info.columns.size() - i]);
-        }
-
-        ignoreErrors =
-            bindData->config.getOption(CopyConstants::IGNORE_ERRORS_OPTION_NAME, ignoreErrors);
-    }
+    expression_vector warningDataExprs = boundSource->getWarningColumns();
+    const bool ignoreErrors = boundSource->getIgnoreErrorsOption();
     auto columns = boundSource->getColumns();
     auto offset = expressionBinder.createVariableExpression(LogicalType::INT64(),
         std::string(InternalKeyword::ROW_OFFSET));