--- a/src/include/binder/copy/index_look_up_info.h
+++ b/src/include/binder/copy/index_look_up_info.h
@@ -9,12 +9,14 @@ struct IndexLookupInfo {
     common::table_id_t nodeTableID;
     std::shared_ptr<binder::Expression> offset; // output
     std::shared_ptr<binder::Expression> key;    // input
+    std::vector<std::shared_ptr<binder::Expression>> warningData;
 
     IndexLookupInfo(common::table_id_t nodeTableID, std::shared_ptr<binder::Expression> offset,
-        std::shared_ptr<binder::Expression> key)
-        : nodeTableID{nodeTableID}, offset{std::move(offset)}, key{std::move(key)} {}
-    IndexLookupInfo(const IndexLookupInfo& other)
-        : nodeTableID{other.nodeTableID}, offset{other.offset}, key{other.key} {}
+        std::shared_ptr<binder::Expression> key,
+        std::vector<std::shared_ptr<binder::Expression>> warningData = {})
+        : nodeTableID{nodeTableID}, offset{std::move(offset)}, key{std::move(key)},
+          warningData(std::move(warningData)) {}
+    IndexLookupInfo(const IndexLookupInfo& other) = default;
 };
 
 } // namespace binder