--- a/src/include/binder/copy/index_look_up_info.h
+++ b/src/include/binder/copy/index_look_up_info.h
@@ -9,13 +9,12 @@ struct IndexLookupInfo {
     common::table_id_t nodeTableID;
     std::shared_ptr<binder::Expression> offset; // output
     std::shared_ptr<binder::Expression> key;    // input
-    std::vector<std::shared_ptr<binder::Expression>> warningData;
+    binder::expression_vector warningExprs;
 
     IndexLookupInfo(common::table_id_t nodeTableID, std::shared_ptr<binder::Expression> offset,
-        std::shared_ptr<binder::Expression> key,
-        std::vector<std::shared_ptr<binder::Expression>> warningData = {})
+        std::shared_ptr<binder::Expression> key, binder::expression_vector warningExprs = {})
         : nodeTableID{nodeTableID}, offset{std::move(offset)}, key{std::move(key)},
-          warningData(std::move(warningData)) {}
+          warningExprs(std::move(warningExprs)) {}
     IndexLookupInfo(const IndexLookupInfo& other) = default;
 };