… dataset/copy-fault-tests/invalid-row/eWatches.csv
… src/binder/CMakeLists.txt
… src/include/processor/operator/persistent/batch_insert.h
… src/include/processor/operator/persistent/batch_insert_error_handler.h
… src/include/processor/operator/persistent/node_batch_insert.h
… src/include/processor/operator/persistent/node_batch_insert_error_handler.h
… src/include/processor/operator/persistent/rel_batch_insert.h
… src/include/processor/warning_context.h
… src/planner/plan/plan_copy.cpp
… src/processor/map/map_copy_from.cpp
… src/processor/map/map_index_scan_node.cpp
… src/processor/operator/index_lookup.cpp
… src/processor/operator/partitioner.cpp
… src/processor/operator/persistent/CMakeLists.txt
… src/processor/operator/persistent/batch_insert_error_handler.cpp
… src/processor/operator/persistent/index_builder.cpp
… src/processor/operator/persistent/node_batch_insert.cpp
… src/processor/operator/persistent/node_batch_insert_error_handler.cpp
… src/processor/operator/table_function_call.cpp
… src/processor/warning_context.cpp
… test/test_files/exceptions/copy/ignore_invalid_row.test
… test/test_files/exceptions/copy/invalid_row.test
… test/test_files/exceptions/copy/null_pk.test