# Description

Partially Addresses #1872

Index lookup failures will now be ignored if `IGNORE_ERRORS=true` for `COPY_FROM` and `LOAD_FROM`. Any row in a rel that has either PK index reference lookup fail will be ignored (won't be added to the rel table). Additionally, a warning containing the error message (and warning source data only if the source is a CSV file) will be cached in the warning table.

# Contributor agreement

- [ ] I have read and agree to the [Contributor Agreement](https://github.com/kuzudb/kuzu/blob/master/CLA.md).