@royi-luo 2024-09-17
Some quick benchmarks:

## Dataset
Schema:
```
CREATE NODE TABLE person(ID UINT32, PRIMARY KEY(ID));
CREATE REL TABLE knows(FROM person TO person, MANY_MANY);
```
Details: 10M nodes in `person`, 10M rels in `knows`
Copy query: `COPY knows FROM "/home/kuzu/test/dataset/missing-pk-rels.csv"(IGNORE_ERRORS=false, DELIM=" ");`

## Results
Branch|`IGNORE_ERRORS`|Error %|Copy Time
----|----|----|----
master|false|0%|1581ms
new|false|0%|1582ms
new|true|0%|2437ms
new|true|10%|2288ms
new|true|100%|1352ms

As the % of errors increases the copy time actually increases since for index lookup errors we avoid copying erroneous rows to the rel table (and so we save the copy time)